Split optional-dep tests into separate file for CI compatibility
CI runs tests twice: first without pyarrow/tomli-w, then with.
Module-level imports of pyarrow caused collection errors in the
first run. Move tests requiring optional deps into a separate file
guarded by pytest.importorskip so they are cleanly skipped.
